Refactor Group4 reader into stream module
