+++ tests/decode_images.rs
@@ -676,5 +676,5 @@ fn bytes_gray_f32() {
 #[test]
 #[cfg(feature = "fax")]
 fn test_fax4() {
-    test_image_sum_u8("fax4.tiff", ColorType::Gray(1), 62384985);
+    test_image_sum_u8("fax4.tiff", ColorType::Gray(1), 7802706);
 }
